New module 'traces' to analyse / process HiSPARC traces.  Added it to the documentation.
Fixed a bug which disabled ctrl-c permanently after download.

Detect and handle a few eventwarehouse bugs due to strange signed int ->
unsigned int -> double mismatches.  For example, 4294966297 is actually
2**32 - 999, which means that it is an unsigned interpretation of a signed
number.  This is then stored in a MySQL double.  The value should be -999
(error in baseline calculation), so restore that value.