Added first version of event generator for front-to-back testing.
Fixed a glaring bug which poisoned my dataset in legacy.py.  OOTI
introduced 9000 ns uniform randomness to the timestamps of events.  This
showed in my KASCADE analysis and ended up in the pamflet.  We've puzzled
over non-gaussian clock jitter with a width of about 10 us for many, many
hours.  This triggered the testing requirement.