Don't choke on rerunning simulations
Both store_aires_data and detector_sim are now resilient against
rerunning.  If previous data already exists, ignore it.  If data can't be
found, graciously return.  This enables one to add simulations to the code
and rerun the entire script while only the new simulations get run.  This
makes it possible to maintain a script that will rebuild everything from
scratch if the data file is missing or removed.