Fix to make clusters picklable
Class variables vs instance variables... sigh.  Anyway, I had a list in
the class definition, so that was a class variable.  When I append to that
list, it is *still* a class variable.  Worse, I'm *really appending to the
class list, so the new values are shared between all instances!!!*

Only instance variables are pickled, so it really breaks.  Fix: let an
instance assign to that variable and it is now an instance variable.