Behavior test form BaseSimulation
Created a nice behavior test for the BaseSimulation class.  It performs an
actual simulation and compares it to a previous one, column by column.
The observables and coincidences need to be the same.

To suppress progressbar output, explicitly initialize it with
fd=sys.stderr.  Unfortunately, the progressbar source initializes the
default value (also sys.stderr) at import time, making it impossible to
redirect.