Get/set coordinates for cluster and station
A station now has knowledge of the containing cluster and the position and
angle of a cluster can be set with set_coordinates and read with
get_coordinates.  Station now has a new method new_get_coordinates where
you no longer need to explicitly give the cluster's position.  That
position is read from cluster.get_coordinates instead.  That means that in
order to perform a simulation where you move the cluster around
(GroundParticleSimulation) you need to explicitly set the cluster
coordinates;  there is no longer a need to pass cluster coordinates
around.