BaseLdfSimulation correct
The class tracks the number of rows in the observables table by using a
private variable which gets updated after each row.append().  At the end
of the simulation, the table is flush()ed and the value of the private
variable must be equal to the actual number of rows on disk.

Since we also test the correct use of this variable in a unit test, the
acceptance test is dropped (it was not yet implemented).