GroundParticlesSimulation: transform coordinates
The GroundParticlesSimulation works by taking a static shower for each
event place a cluster on a different location and rotated randomly around
the cluster center.  However, we'd like to think the cluster is static,
and there are different showers randomly falling on and around the
cluster.

This coordinate transformation takes care of that.

(Tests not fully updated yet!!!)