New reconstruction efficiency analysis script
Landau PDF pre-computes values and interpolates between them.  Massive
speedup!  Reconstruction efficiency first does a exponential fit to the
gamma part of the pulseheight spectrum, then does a landau/gauss fit to
the charged particle part of the spectrum, and finally uses the determined
parameters as starting values for a complete spectrum fit.