Added constrained optimizer
Not very succesful.  While it does keep the optimization from 'running
away', it is not as accurate.  I used the truncated Newton optimization,
which needs to approximate a gradient (as I don't have one).  It is less
accurate then the Nealder-Mead simplex method, which I use for the
unconstrained optimizer.