Restructured simulations
Moved the gauss/poisson detector signal simulation to BaseSimulation.
Removed write_observables override in BaseLdfSimulation.  Removed
num_particles and detector timings logic from BaseSimulation
store_observables.  It is now the responsibility of the simulation to
correctly simulate detector signals.

The point of all this is that the GroundParticles simulation can now use
Poisson and Gauss too.

Had to break one test.  Overreliance on sentinels is hard to maintain.  No
time to fix that now.