Wrote utility class for working with GPS time
The class accepts datetime instances as well as timestamps and interprets
them as GPS time.  You can then use the class instance to quickly convert
to a string, a timestamp or a datetime instance.

This reduces much of the hassle of working with the HiSPARC data.  More
importantly, it reduces the chance of confusing local time with UTC or GPS
time.