FIX: particle arrival times now use baseline
The particle arrival time code used the first 100 samples of the trace to
determine the baseline.  These samples are cut off by the 'reduce data'
feature in the LabVIEW DAQ.  The implementation of 'reduce 'data' is
faulty since it should leave 100 samples intact before the start of the
trace.  The new code uses the baseline as determined by the DAQ and does
not use the first 100 samples.  Closes #3.