Implement functions to prepare output tables
One function creates the coincidences group and the
containing tables; coincidences, c_index and s_index.
The coincidences group also gets the cluster as attribute.

Another function creates the events tables for each station
First a cluster_simulations group is created,
this contains the station groups which contain the events tables.

Finally a function populates the s_index table with the paths to the
station groups.