Implement function to store station observables
Currently does not work, unknown what format station_observables will
be.

Sets event_id to the row number (0-based),
the rest of values should come from station_observables.
Returns the event_index for the c_index

The run now loops over an enumerated list of stations.
The station_id is used for finding the station group and the c_index.