Add station_number property to Station class
A station_number can be given (optional) when adding a station to a
cluster.
If no number is given the station_id will be used for the
station_number.
Add a function to get a station based on station_number.
station_id is now 0-based, for consistency.

(cherry picked from commit acdffa7923603a09bcca872982fe30b20d89edfb)