Add two new classes for settings up clusters
HiSPARCStations for any combination of stations from the network.
The latest gps coordinates and the detector number is retrieved.
Detector positions are not yet taken into account (even for Science
Park).

HiSPARCNetwork creates a cluster objects for the entire HiSPARC Network.
Stations that have never uploaded a configuration will be skipped.