Fix azimuth in output
Corsika has a very interesting definition of azimuth,
it defines azimuth as the direction the shower is pointing to.
Moreover, it uses degrees in the input file and radians in the output.
It also allows values between -2π and 2π.

This is now fixed, in corsika.blocks.
π is added to the Corsika azimuth to fix the direction,
and then fixed to be between -π and π.
So regenerate your h5 from Corsika data!

simulations.groundparticles now correctly adds alpha and azimuth as the
azimuth of the shower (also made to fit in (-π, π]).
And alpha is now correctly passed to the cluster preparation function.