Skip ProcessEventsTests if no test datafile exists
The PE-testdata.h5 file is >50 MB.
A smaller sample should be created for this test.
Small enough that it can be added to the repository,
once that is done and the tests have been adapted,
it can be removed from the .gitignore to be included.

For now skip tests if the file is not available,
otherwise tests on Travis will take a lot longer.