Fix corsika run_header conversion between height and thickness
Add info to docstring about where the model is specified in the CORSIKA
manual.
These functions were missing self. in front of [abc]_atomospheric.
These lists are now stored in temporary variables to improve the
overview.
Also the wrong values were used for the various atmospheric layers.
And some of the formulas were missing parameters (even though they may
always be 1)
Determine layer boundaries in thickness by using the
height_to_thickness conversion,
this ensures that it is correct for the used atmospheric parameters.
Add a test to see if a round trip between these functions gives back
the input (for each layer).