Some improvements to FitAlgorithm
FitAlgorithm turned out to perform badly with integer starting values
(or with starting values equal to 0.0). It also seemed to require
specified values for the options ftol, eps and maxiter. Finally, the
code is a little bit enlarged to find a possible second zenith value.
The handling of the theta's is made similar to DirectAlgorithm3D.