Update the docs
Move docs about submodules into folders, to keep things a bit organized.
Fix some references to modules, classes and functions.
Update the tutorial to use the esd module instead of the publicdb.
This makes the tutorial a lot faster and user friendly.
Fixed a bug in CoincidencesESD concerning the cluster attribute.