Show discrete reconstruction solutions
For a station (4 detectors), show the possible arrival directions for
arrival times in discrete steps of 2.5 ns.
All possible combinations of arrival times with steps of 2.5 ns are
created and reconstructed.
Results (zenith, azimuth) are shown for the different combinations of 3
detectors.
And the layout of the station is shown.