Add station offsets for coincidence reconstruction
With several caveats..
Currently uses 501 as reference station.
Therefore, this will not work for stations far from Science Park,
so for each (sub)cluster a different reference should be chosen.
Also the binning may not be optimal yet.
And station offsets are better determined when there is more
than one day of coincidence data available..