New find MPV algorithm
This algorithm first fits the data with a power law and subtracts it
from the data to bring out the MIP peak. Then, an estimator for the
smoothness of the data is used to check whether there actually is a MIP
peak in the data. If so, use the fast algorithm on the reduced data to
locate it.

Thanks to Hans for suggesting the power law (instead of an exponential
function) and for coming up with a good estimator for the smoothness.