Fix coincidence reconstructions
The ids in the reconstructions of coincidences did not match
the ids of the coincidences they belong too, because only
coincidences with n >= 2 events were returned by cq.all_events.
This caused self.coincidences and the reconstructed self.theta
(etc) to have different lengths.
Also correct the conversion from zenith to latitude.