Add extra step to CORSIKA data conversion
In addition to creating a cs index the table will now be sorted by x.
This speeds up simulations by more than a magnitude (upto 100x).
The 'old' corsika.h5 is now created as a temporary/intermediate file.
A sorted version is then created, after which the temp file is removed.
So during the conversion a bit more free disc space is required.
Compression was considered but did not show any advantages in tests.

The `copy_and_sort_node` is in this case equal to:
`ptrepack --sortby x --propindexes old.h5 new.h5`

Thanks to @tomkooij for investigating!