Improve shifts for Coincidences
Both _search_coincidences _retrieve_timestamps use the same units for
shifts.
Fix the docstrings, mention the units for shifts and remove
non-existent param.
Rename parameter to more correctly describe what it is.
Update the tests to actually test the values.
Comparing uint64 to an int is not accurate to nanoseconds.
Include a test that ideally would fail, but does not.