Add warnings to CoincidenceQueries
It should not happen, but in case c_index does not match
s_index ids for station paths wrong station paths may be used.
In case of downloaded coincidences for a specific set of stations
the group may be missing, since s_index contains a group for all
stations.
Clarify an exception for esd.download_coincidences.