Remove cluster from Coincidences(ESD)
The cluster object was now unused for Coincidences.
For CoincidencesESD it has been replaced by a station_numbers keyword.
The only thing the cluster object was used for was to extract station
numbers.
This is easier to do from the list.
Added exception in case the number of numbers does not match the number
of groups.
Updated test data and docs.