Add support for the end status line for downloading
On a successful download the last line should be a non-empty comment
line.
If the last line is indeed a non-empty comment, do nothing.
If it is not raise an exception.
If the last line was an empty comment only the header was downloaded.
If the last line is data report the date/time of that data.