Implement force_fresh and force_stale for API
These options replace the allow_stale option.
If both are false allow_stale is implied.
If either a true either local data will be used,
or data must be retrieved from the public database.
Move assignment of info to separate lazy method.
This speeds up initialisation of Station class.
It does not (yet) check if it is a valid station.