Fix flushing last data
When (down)loading data the last data/coincidence was not flushed.
It would automatically be flushed when the file is closed.
However, if the file is not closed between subsequent downloads to
the same data file and tables the last event id will be repeated.
This is because the length of the table will be one short,
because the last row was not yet flushed.
Thanks @SBeijen for helping find this bug.