Remove warning when station groups are missing
When downloading coincidences from a subset of the network, the s_index
will have paths to all station groups currently in the network. However
only the stations actually in the downloaded coincidences will have event
tables in those station groups. This is not a problem. The warning can
thus be removed. When the events are requested (in reconstructions) a warning
will be thrown by _get_event().